culminando la tabla cupones
crud finalizado de la tabla cupones, que muestra los cupones de cada cliente

+++ database/migrations/2023_07_15_185301_create_cupones_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         
         Schema::create('cupones', function (Blueprint $table) {
             $table->id('id_cupon');
-            $table->double('descuento');
+            $table->decimal('descuento');
             $table->date('fecha_expiracion');
             $table->string('codigo_cupon');
             $table->foreignId('id_cliente')->constrained('clientes', 'id_cliente');
